+++ src/test/java/io/github/astrapi69/lang/ClassExtensionsTest.java
@@ -50,7 +50,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import io.github.astrapi69.classes.inner.OuterClass;
-import io.github.astrapi69.collections.array.ArrayFactory;
+import io.github.astrapi69.collection.array.ArrayFactory;
 import io.github.astrapi69.runtime.compiler.JavaSourceCompiler;
 import io.github.astrapi69.test.object.Member;
 import io.github.astrapi69.test.object.Person;
@@ -86,6 +86,28 @@ public class ClassExtensionsTest
 	}
 
 	/**
+	 * Test method for {@link ClassExtensions#isInstantiable(Class)}
+	 */
+	@Test
+	public void testIsInstantiable()
+	{
+		boolean expected;
+		boolean actual;
+
+		actual = ClassExtensions.isInstantiable(Brand.class);
+		expected = false;
+		assertEquals(expected, actual);
+
+		actual = ClassExtensions.isInstantiable(Integer.class);
+		assertEquals(expected, actual);
+
+		actual = ClassExtensions.isInstantiable(Person.class);
+		expected = true;
+		assertEquals(expected, actual);
+
+	}
+
+	/**
 	 * Test method for {@link ClassExtensions#forName(String)}
 	 *
 	 * @throws ClassNotFoundException
@@ -94,10 +116,18 @@ public class ClassExtensionsTest
 	@Test
 	public void testForName() throws ClassNotFoundException
 	{
-		final Class<?> expected = this.getClass();
-		final String classname = "io.github.astrapi69.lang.ClassExtensionsTest";
-		final Class<?> actual = ClassExtensions.forName(classname);
+		Class<?> expected;
+		Class<?> actual;
+		String classname;
 
+		classname = "io.github.astrapi69.lang.ClassExtensionsTest";
+		actual = ClassExtensions.forName(classname);
+		expected = this.getClass();
+		assertEquals(expected, actual);
+
+		classname = "io.github.astrapi69.test.object.Person";
+		actual = ClassExtensions.forName(classname);
+		expected = Person.class;
 		assertEquals(expected, actual);
 	}
 
@@ -110,15 +140,12 @@ public class ClassExtensionsTest
 	@Test
 	public void testGetRunningJarFile() throws URISyntaxException
 	{
-		String actual;
-		String expected;
-
 		File runningJarFile = ClassExtensions.getRunningJarFile(Person.class);
 		assertNotNull(runningJarFile);
 
 		String absolutePath = runningJarFile.getAbsolutePath();
 		assertTrue(absolutePath.endsWith(".jar"));
-		assertTrue(absolutePath.contains("test-objects"));
+		assertTrue(absolutePath.contains("test-object"));
 	}
 
 	/**
@@ -333,7 +360,6 @@ public class ClassExtensionsTest
 			"FooRunnable", source);
 
 		actual = ClassExtensions.getClassType(clazz);
-		expected = ClassType.DEFAULT;
 		assertEquals(expected, actual);
 
 		actual = ClassExtensions.getClassType(Brand.class);
@@ -753,11 +779,9 @@ public class ClassExtensionsTest
 
 
 		actual = ClassExtensions.isDerivate(null, null);
-		expected = true;
 		assertEquals(expected, actual);
 
 		actual = ClassExtensions.isDerivate(null, ClassExtensions.getClassLoader());
-		expected = true;
 		assertEquals(expected, actual);
 
 		actual = ClassExtensions.isDerivate(ClassExtensions.getClassLoader(), null);
@@ -855,7 +879,6 @@ public class ClassExtensionsTest
 		InvocationHandler invocationHandler = new InvocationHandlerHandler<>(bla);
 		Foo proxy = (Foo)Proxy.newProxyInstance(ClassExtensions.getClassLoader(),
 			new Class[] { Foo.class }, invocationHandler);
-		expected = true;
 		actual = ClassExtensions.isProxy(proxy.getClass());
 		assertEquals(expected, actual);
 	}
@@ -894,8 +917,7 @@ public class ClassExtensionsTest
 		public Object invoke(Object proxy, Method method, Object[] args)
 			throws IllegalAccessException, IllegalArgumentException, InvocationTargetException
 		{
-			Object object = method.invoke(original, args);
-			return object;
+			return method.invoke(original, args);
 		}
 	}
 
@@ -915,8 +937,7 @@ public class ClassExtensionsTest
 		public Object intercept(Object o, Method method, Object[] args, MethodProxy methodProxy)
 			throws Throwable
 		{
-			Object object = method.invoke(origin, args);
-			return object;
+			return method.invoke(origin, args);
 		}
 	}