+++ src/test/java/io/github/astrapi69/lang/model/PropertiesKeyAndParametersTest.java
@@ -30,7 +30,7 @@ import org.meanbean.test.Configuration;
 import org.meanbean.test.ConfigurationBuilder;
 import org.testng.annotations.Test;
 
-import io.github.astrapi69.collections.array.ArrayFactory;
+import io.github.astrapi69.collection.array.ArrayFactory;
 
 /**
  * The unit test class for the class {@link PropertiesKeyAndParameters}.