+++ src/test/java/io/github/astrapi69/lang/thread/ThreadFieldDataBeanTest.java
@@ -29,7 +29,7 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import io.github.astrapi69.collections.map.MapFactory;
+import io.github.astrapi69.collection.map.MapFactory;
 
 /**
  * The unit test class for the class {@link ThreadFieldDataBean}.